remove(ng): extract rectangles
