+++ app/javascript/angular/code/services/google/map.js
@@ -7,7 +7,6 @@ angular
     "$cookieStore",
     "$rootScope",
     "digester",
-    "rectangles",
     "googleMaps",
     "$window",
     map
