+++ app/javascript/angular/code/services/update_crowd_map_layer.js
@@ -8,7 +8,6 @@ angular
     "params",
     "utils",
     "infoWindow",
-    "rectangles",
     "$window",
     updateCrowdMapLayer
   ]);
